+++ assa/src/App.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import {Route, Redirect, Switch} from 'react-router-dom';
 import {ConnectedRouter} from 'connected-react-router';
-import Login from './containers/login/Login';
-import Main from './containers/main/Main'
+import Login from './containers/Login/Login';
+import Main from './containers/Main/Main'
 import './App.css';
 function App(props) {
     return (
